\name{kap4}
\alias{kap4}
\docType{data}
\title{
Datasett til kapittel 4
}
\description{
Datasett for å løse oppgavene i kapittel 4.
}
\usage{data("kap4")}
\format{
  Et datasett med 8 observasjoner på følgende 4 variabler.
  \describe{
    \item{\code{parti}}{en faktor med følgende nivåer \code{Ap}; \code{FrP}; \code{H}; \code{KrF}; \code{MDG}; \code{SV}; \code{Sp}; \code{V}}
    \item{\code{hoyre}}{en numerisk variabel. Partiers høyreorientering (fra 0 til 10)}
    \item{\code{omfordeling}}{en numerisk variabel. Partiers preferanse for omfordeling (fra 0 til 10)}
    \item{\code{viktighet}}{en numerisk variabel. Viktigheten av spørsmål knyttet til økonomiske vestre-høyreorientering}
  }
}

\references{
Det originale datasettet er trukket fra:

Ryan Bakker, Erica Edwards, Liesbet Hooghe, Seth Jolly, Gary Marks, Jonathan Polk, Jan Rovny, Marco Steenbergen, and Milada Vachudova (2015) "2014 Chapel Hill Expert Survey." Version 2015.1. Chapel Hill, NC: University of North Carolina, Chapel Hill.

Data: \url{http://chesdata.eu}.

Kodebok: \url{https://www.chesdata.eu/2014-chapel-hill-expert-survey}
}
\examples{
data(kap4)
}
\keyword{datasets}
