\name{kap5}
\alias{kap5}
\docType{data}
\title{Datasett for kapittel 5
}
\description{
Datasett med norske respondenter fra 2014. Dataene er tilrettelagte fra Den europeiske samfunnsundersøkelsens 7. runde. ESS-variabler er indikert i parentes.
}
\usage{data("kap5")}
\format{
  Et datasett med 1436 observasjoner av følgende 14 variabler.
  \describe{
    \item{\code{Parti}}{en kategorisk variabel} Hvilket parti stemte du på ved forrige stortingsvalg?
    \item{\code{Inntekt}}{en tellevariabel}
    \item{\code{Utdanning}}{en tellevariabel}
    \item{\code{Kvinne}}{en binær variabel}
    \item{\code{Innvandrer}}{en binær variabel}
    \item{\code{Skepsis}}{en tellevariabel}
    \item{\code{KultSkepsis}}{en tellevariabel}
    \item{\code{TaJobber}}{en tellevariabel}
    \item{\code{TaSkatt}}{en tellevariabel}
    \item{\code{Belastning}}{en tellevariabel}
    \item{\code{Arbeidsledig}}{en binær variabel}
    \item{\code{Begrenset}}{en binær variabel}
    \item{\code{IngenKontrakt}}{en binær variabel}
    \item{\code{Prekaritet}}{en binær variabel}
  }
}
\source{
ESS Round 7: European Social Survey Round 7 Data (2014). Data file edition 2.1. NSD - Norwegian Centre for Research Data, Norway - Data Archive and distributor of ESS data for ESS ERIC.

Tilgjengelig på: \url{http://www.europeansocialsurvey.org/data/download.html?r=7}
}
\examples{
data(kap5)
}
\keyword{datasets}
