\name{laerdegR}
\alias{laerdegR}

\title{Støttepakke til Lær deg R}

\description{
Denne R-pakken er en nettressurs beregnet på boka Lær deg R (2019).
}

\format{
  Bruk pakken aktivt når du leser boka.
  \describe{
    \item{\strong{Last inn eksmpeldata}}{Lek sammen med meg! Den beste måten å lese boka på, er ved tastaturet. Hvert kapittel er utstyrt med et eksempeldatasett som er oppkalt etter boka. Slik er \code{kap7} datasettet til kapittel 7.
      }
  \item{\strong{Løs oppgaver}}{Hvert kapittel har øvelsesoppgaver med løsningsforslag liggende i denne pakken. Slik er \code{oppg7.1()} løsningsforslaget til oppgave 7.1.
      }
  \item{\strong{Sjekk ut figurene}}{Kodene til figurene i boka ligger som funksjone i denne pakken. Da kan du se hvordan jeg har tilpasset dem. Slik er \code{fig8.1()} en funksjon som oppretter figur 8.1.
      }

}
}

\author{
Silje Synnøve Lyder Hermansen
}

