\name{oppg2.2}
\alias{oppg2.2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Oppgave 2.2
}
\description{
Regn ut medianen for vektoren \code{hoyre}. Finnes det noe parti med denne eksakte verdien av høyreorientering?
}
\usage{
oppg2.2()
}

\author{
Silje Synnøve Lyder Hermansen
}

\examples{
oppg2.2()
oppg2.2
}
