\name{oppg2.3}
\alias{oppg2.3}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Oppgave 2.3
}
\description{
Du finner en variabel som indikerer viktigheten av spørsmål knyttet til økonomisk vestre-høyreorientering for hvert parti i datatabellen i kapittel 2. Legg variabelen til datasettet.

Hvilket parti har mest å vinne på at velgere vektlegger økonomiske spørsmål?

}
\usage{
oppg2.3()
}

\author{
Silje Synnøve Lyder Hermansen
}

\examples{
oppg2.3()
oppg2.3
}
