\name{oppg2.4}
\alias{oppg2.4}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Oppgave 2.4
}
\description{
Opprett en ny variabel kalt \code{hoyre.kat}.

Omkod så variabelen til tre kategorier på grunnlag av \code{hoyre}: ``Venstre'', ``Sentrum'' og ``Hoyre''. Gjør variabelen om til en faktor.

Hvilke partier befinner seg på høyresiden?

Bruk indeksering aktivt i denne oppgaven!
}
\usage{
oppg2.4()
oppg2.4
}

\author{
Silje Synnøve Lyder Hermansen
}

\examples{
oppg2.4()
oppg2.4
}
