\name{oppg3.1}
\alias{oppg3.1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Oppgave 3.1
}
\description{
Tegn et spredningsdiagram for å gi en grafisk framstilling av korrelasjonen mellom partiers høyreorientering  (ved hjelp av hoyre) og preferanser for økonomisk omfordeling (ved hjelp av omfordeling).

Bruk plot()-funksjonen direkte.
}
\usage{
oppg3.1()
oppg3.1
}

\references{
Ryan Bakker, Erica Edwards, Liesbet Hooghe, Seth Jolly, Gary Marks, Jonathan Polk, Jan Rovny, Marco Steenbergen, and Milada Vachudova (2015) "2014 Chapel Hill Expert Survey." Version 2015.1. Chapel Hill, NC: University of North Carolina, Chapel Hill.

Data: \url{http://chesdata.eu}.

Kodebok: \url{https://www.chesdata.eu/2014-chapel-hill-expert-survey}
}
\author{
Silje Synnøve Lyder Hermansen
}
