\name{oppg3.3}
\alias{oppg3.3}
\title{
Oppgave 3.3
}
\description{
Tegn det samme diagrammet ved å legge inn observasjonene med partinavn og gi dem en relevant farge. Start med å opprette et tomt koordinat system. Legg så inn partinavn og spesifiser farge.

Om partinavnene er definert som en kategorisk variabel (faktor), må du redefinere variabelen som bokstaver ved hjelp av \code{as.character()}.

}
\usage{
oppg3.3()
oppg3.3
}

\author{
Silje Synnøve Lyder Hermansen
}


\seealso{
oppg3.2
}
\examples{
data(kap3)
oppg3.3()
}
