\name{oppg3.4}
\alias{oppg3.4}
\title{
Oppgave 3.4
}
\description{
Bestem størrelsen på partinavnene etter hvor viktig den økonomiske høyre-venstreaksen (\code{kap3$viktighet}) er for hvert av partiene ved å bruke \code{cex=}. Om partinavnene blir for store, kan du redusere vektene ved å dele hele variabelen på en konstant (et tall, f.eks fire).
}
\usage{
oppg3.4()
oppg3.4
}

\author{
Silje Synnøve Lyder Hermansen
}

\seealso{
oppg3.2;
oppg3.3
}
\examples{
data(kap3)
oppg3.4()
}
