\name{oppg4.3}
\alias{oppg4.3}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Oppgave 4.3
}
\description{
Last ned og innstaller R-pakken \code{datasets}.

Åpne og inspiser datasettet for biler (\code{cars}).

Inspiser hjelpesiden for dette datasettet.
}
\usage{
oppg4.3()
oppg4.3
}

\author{
Silje Synnøve Lyder Hermansen
}

\examples{
oppg4.3()
oppg4.3
}
