\name{oppg4.4}
\alias{oppg4.4}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Oppgave 4.4
}
\description{
Eksporter datasettet \code{kap4} ved hjelp av \code{write.table()}. Separer kolonnene med semikolon og fjerne radnumrene.

}
\usage{
oppg4.4()
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Silje Synnøve Lyder Hermansen
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
oppg4.4()
oppg4.4
}
