\name{oppg5.1}
\alias{oppg5.1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Oppgave 5.1
}
\description{
Ta en titt på variablene i datasettet. Identifiser to variabler som er binære, kontinuerlige og estår av heltall (tall uten desimaler). Hvilke grafiske fremstillinger er de mest hensiktsmessige for å utforske den univariate fordelingen til disse variablene?

}
\usage{
oppg5.1()
}

\author{
Silje Synnøve Lyder Hermansen
}

\examples{
oppg5.1()
}
